Move migration code into its own class
